+++ CHANGES.md
@@ -1,5 +1,6 @@
 Version 1.1.2-dev
 -----------------
+- On PyTorch 2.12+, enable `expandable_segments` in `startup` also for ROCm.
 
 
 Version 1.1.1 [29 Jun 2026]